requirements.txt
langchain
langchain-community
langchain-openai
chromadb
python-dotenv